reddit_bot/requirements.txt
praw==7.8.1
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.0
colorama==0.4.6
Pillow==10.1.0
